removed duplicates, account mocks for devlive
