+++ __mocks__/wallet0.json
@@ -1,0 +1,1 @@
+{"address":"273b2d843b0a8bad5ac4d9fd18719998a7c33bcc","id":"face79bd-adbb-4c97-b9be-c07af015cf06","version":3,"Crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"c27befca7b77aa4e451ac28e9a7e4744"},"ciphertext":"4a71931c314b5328eeaa5f02c49febe44cf4feaf912959f57663597433293f2a","kdf":"scrypt","kdfparams":{"salt":"df7be65ee0810c6bf08106fd010c9ae313988d94f4c969bcd4d29897159625a2","n":131072,"dklen":32,"p":1,"r":8},"mac":"7e63c690da00705570e0c39fadc64fb654954ca086e3c7185c847f33dba527aa"},"x-ethers":{"client":"ethers.js","gethFilename":"UTC--2021-07-12T20-13-06.0Z--273b2d843b0a8bad5ac4d9fd18719998a7c33bcc","mnemonicCounter":"8b3940c3db921e0d90c56a40ad3997ad","mnemonicCiphertext":"b8b4fadf850466c8fec67943a1faea93","path":"m/44'/60'/0'/0/0","version":"0.1"}}
