+++ __mocks__/wallet1.json
@@ -1,0 +1,1 @@
+{"address":"9f982f897251c2a97f847e0ca3966f9645ea5698","id":"a0801dc8-f259-47d5-8c3b-95979eb5718b","version":3,"Crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"051bd98dadb3470439b8593ac5e6d78e"},"ciphertext":"dbf52849d40f1692f63f48e62678f667aa274ead8c3595fdd02f437c133350d4","kdf":"scrypt","kdfparams":{"salt":"26f44169f8cd476e4d43987214d216469e4116046b886803f5d0ea980b5a2cc9","n":131072,"dklen":32,"p":1,"r":8},"mac":"4a2bddec980be82adca415edbd95926aa5fdfe6bef0f2377cf106ab1499831d7"},"x-ethers":{"client":"ethers.js","gethFilename":"UTC--2021-07-12T20-13-02.0Z--9f982f897251c2a97f847e0ca3966f9645ea5698","mnemonicCounter":"5831268bd2f31ad0c389636855fcda0d","mnemonicCiphertext":"aa6bca54750108855ff8a1a03b857e83","path":"m/44'/60'/0'/0/0","version":"0.1"}}
