+++ __mocks__/wallet2.json
@@ -1,0 +1,1 @@
+{"address":"7d6ea478d033ee167d3e2abb747b4416d5728a64","id":"3dbdfde6-39ef-494e-8bbf-e2bdc584bf3e","version":3,"Crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"c58a92843d7c6874b4d9b7512cef6ee2"},"ciphertext":"85f734fd06e78aca60488fc045feca871693511a70c77c61f7143c247af12890","kdf":"scrypt","kdfparams":{"salt":"2eb53367d6cca1da4e30753192e3cda775bcf4990c38ae1518e3fc9b64d9598c","n":131072,"dklen":32,"p":1,"r":8},"mac":"ee89f2cc3c9bc9be08df08fd2ae2c89c15881a71a082d9c1f71b02bfa7058c5a"},"x-ethers":{"client":"ethers.js","gethFilename":"UTC--2021-07-12T20-13-05.0Z--7d6ea478d033ee167d3e2abb747b4416d5728a64","mnemonicCounter":"839270a24db870366a4194319d7de011","mnemonicCiphertext":"c1a5643f023c44591db230da53972b68","path":"m/44'/60'/0'/0/0","version":"0.1"}}
