+++ __mocks__/wallets.ts
@@ -1,0 +1,28 @@
+import wallet0 from './wallet0.json';
+import wallet1 from './wallet1.json';
+import wallet2 from './wallet2.json';
+
+// Wallets returned by GDrive (either mocked googleapis or broker-service directly)
+export const wallets = [
+  {
+    id: "0",
+    originalFilename: `wallet0.wallet`,
+    name: "wallet0",
+    type: "Type?",
+    wallet: JSON.stringify(wallet0),
+  },
+  {
+    id: "1",
+    originalFilename: `wallet1.wallet`,
+    name: "wallet1",
+    type: "Type?",
+    wallet: JSON.stringify(wallet1),
+  },
+  {
+    id: "2",
+    originalFilename: `wallet2.wallet`,
+    name: "wallet2",
+    type: "Type?",
+    wallet: JSON.stringify(wallet2),
+  },
+];
